samples/ml/airflow/airflow/requirements.txt
apache-airflow==2.10.2
apache-airflow-providers-snowflake==5.8.0